requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio[asyncio]>=5.11.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pydantic>=2.10.0
email-validator>=2.0.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-generativeai>=0.8.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
slowapi>=0.1.9
